fix broken ccalls and use generic Ints

+++ Examples/NeHe/tut9/tut9.jl
@@ -24,9 +24,9 @@ height    = 480
 STAR_NUM  = 50
 
 type star
-    r::Int64
-    g::Int64
-    b::Int64
+    r::Int
+    g::Int
+    b::Int
     dist::Float64
     angle::Float64
 end
